refactor: 메서드 명 일관화 및 패키지 이동
- Entity는 domain -> infrastructure/persistence
- to/from 중복 사용 및 Domain/ClassName 혼용 로직 변경
- 미사용 클래스 제거
- 개행 추가 (컨벤션 통일)
